Move user-config load and persistence into useAppConfig
fetchConfig, patchConfigUIOptions, the config snapshot and the fourteen
watchers that persist a UI option when it changes all move out of App.vue
into one composable. App.vue keeps the two bits that are genuinely its
own: applying the Vuetify theme, and re-measuring the column ceiling after
a thumbnail-size change, which the composable now asks for through a hook.

App.vue drops from 2455 to 2051 lines.
