+++ frontend/src/App.vue
@@ -4,7 +4,6 @@ import {
   nextTick,
   onBeforeUnmount,
   onMounted,
-  reactive,
   ref,
   watch,
 } from "vue";
@@ -17,7 +16,7 @@ import {
   isReadOnly,
   sessionContext,
 } from "./utils/apiClient";
-import { getUserConfig, patchUserConfig } from "./api/config";
+import { patchUserConfig } from "./api/config";
 import { useSelectionStore } from "./stores/useSelectionStore";
 import { useFilterStore } from "./stores/useFilterStore";
 import { useSortStore } from "./stores/useSortStore";
@@ -43,6 +42,7 @@ import {
 } from "./stores/useViewStore";
 import { redoKeyHint, undoKeyHint } from "./utils/shortcutHints";
 import { useGridRealtimeSync } from "./composables/useGridRealtimeSync";
+import { useAppConfig } from "./composables/useAppConfig";
 
 import SideBar from "./components/panels/SideBar.vue";
 import TitleBar from "./components/TitleBar.vue";
@@ -57,11 +57,6 @@ import NoticeHost from "./components/widgets/NoticeHost.vue";
 import { useFloatingBottomInset } from "./composables/useBottomAnchor";
 import { toPx } from "./utils/floatingBottom.js";
 import { isInternalImageDrag } from "./utils/media.js";
-import {
-  clampSizeLevel,
-  nearestSizeLevelForColumns,
-} from "./utils/thumbnailSizes";
-
 const BACKEND_URL = API_BASE_URL;
 
 // --- Stores ---
@@ -129,22 +124,14 @@ const loading = ref(null);
 const error = ref(null);
 
 // --- Config tracking ---
-const configLoaded = ref(false);
-const configLoading = ref(false);
-const configApplying = ref(false);
-const configSnapshot = ref({});
-const config = reactive({
-  sort: "",
-  thumbnail: 256,
-  thumbnail_mode: "square",
-  sidebar_thumbnail_size: 64,
-  show_stars: true,
-  show_face_bboxes: false,
-  show_problem_icon: true,
-  expand_all_stacks: true,
-  date_format: "locale",
-  theme_mode: "light",
-  stack_strictness: 0.92,
+// Loading the user's config and persisting UI options back lives in
+// useAppConfig; App.vue only supplies the layout re-measure that a
+// thumbnail-size change needs.
+const { fetchConfig } = useAppConfig({
+  onThumbnailSizeChanged: () => updateMaxColumns(),
+  onUpdateCheckUndecided: () => {
+    updateCheckDialogOpen.value = true;
+  },
 });
 
 // --- Layout constants ---
@@ -1224,275 +1211,6 @@ function handleUpdateSidebarWidth(value) {
   userPrefsStore.sidebarWidth = nextValue;
 }
 
-async function fetchConfig() {
-  if (isReadOnly.value) {
-    userPrefsStore.themeMode = "dark";
-    userPrefsStore.sidebarThumbnailSize = 32;
-    gridStore.showProblemIcon = true;
-    gridStore.showFaceBboxes = false;
-    gridStore.showStars = true;
-    return;
-  }
-  if (configLoading.value) return;
-  configLoading.value = true;
-  configApplying.value = true;
-  try {
-    const cfg = await getUserConfig();
-    const sortValue = cfg.sort_order ?? cfg.sort;
-    if (typeof sortValue === "string" && sortValue) {
-      sortStore.selectedSort = sortValue;
-    }
-    if (typeof cfg.show_keyboard_hint === "boolean")
-      userPrefsStore.showKeyboardHint = cfg.show_keyboard_hint;
-    if (typeof cfg.show_face_bboxes === "boolean") {
-      gridStore.showFaceBboxes = cfg.show_face_bboxes;
-    }
-    if (typeof cfg.show_problem_icon === "boolean") {
-      gridStore.showProblemIcon = cfg.show_problem_icon;
-    }
-    if (typeof cfg.expand_all_stacks === "boolean") {
-      gridStore.showStacks = cfg.expand_all_stacks;
-    } else if (typeof cfg.show_stacks === "boolean") {
-      gridStore.showStacks = cfg.show_stacks;
-    }
-    if (typeof cfg.compact_mode === "boolean") {
-      gridStore.compactMode = cfg.compact_mode;
-    }
-    if (typeof cfg.sidebar_docked === "boolean") {
-      sidebarStore.setSidebarDocked(cfg.sidebar_docked);
-    }
-    if (typeof cfg.sidebar_pinned === "boolean") {
-      sidebarStore.setSidebarPinned(cfg.sidebar_pinned);
-    }
-    if (typeof cfg.date_format === "string" && cfg.date_format) {
-      userPrefsStore.dateFormat = cfg.date_format;
-    }
-    if (typeof cfg.theme_mode === "string" && cfg.theme_mode) {
-      userPrefsStore.themeMode = cfg.theme_mode;
-    }
-    if (typeof cfg.descending === "boolean") {
-      sortStore.selectedDescending = cfg.descending;
-    }
-    if (typeof cfg.thumbnail_size_level === "number") {
-      gridStore.sizeLevel = clampSizeLevel(cfg.thumbnail_size_level);
-    } else if (typeof cfg.columns === "number") {
-      // Legacy config predating the size ladder: derive the nearest level so
-      // an old install keeps roughly the same tile size after upgrading.
-      gridStore.sizeLevel = nearestSizeLevelForColumns(cfg.columns);
-    }
-    if (cfg.thumbnail_mode === "square" || cfg.thumbnail_mode === "justified") {
-      gridStore.thumbnailMode = cfg.thumbnail_mode;
-    }
-    if (typeof cfg.sidebar_thumbnail_size === "number") {
-      userPrefsStore.sidebarThumbnailSize = cfg.sidebar_thumbnail_size;
-    }
-    if (typeof cfg.sidebar_width === "number") {
-      userPrefsStore.sidebarWidth = cfg.sidebar_width;
-    }
-    if (cfg.stack_strictness != null) {
-      sortStore.stackThreshold = String(cfg.stack_strictness);
-    }
-    config.sort_order = sortValue || sortStore.selectedSort;
-    config.descending = sortStore.selectedDescending;
-    config.columns = gridStore.columns;
-    config.thumbnail_mode = gridStore.thumbnailMode;
-    config.sidebar_thumbnail_size = userPrefsStore.sidebarThumbnailSize;
-    config.sidebar_width = userPrefsStore.sidebarWidth;
-    config.show_stars =
-      typeof cfg.show_stars === "boolean"
-        ? cfg.show_stars
-        : gridStore.showStars;
-    config.show_face_bboxes =
-      typeof cfg.show_face_bboxes === "boolean"
-        ? cfg.show_face_bboxes
-        : gridStore.showFaceBboxes;
-    config.show_problem_icon =
-      typeof cfg.show_problem_icon === "boolean"
-        ? cfg.show_problem_icon
-        : gridStore.showProblemIcon;
-    config.expand_all_stacks =
-      typeof cfg.expand_all_stacks === "boolean"
-        ? cfg.expand_all_stacks
-        : typeof cfg.show_stacks === "boolean"
-          ? cfg.show_stacks
-          : gridStore.showStacks;
-    config.compact_mode =
-      typeof cfg.compact_mode === "boolean"
-        ? cfg.compact_mode
-        : gridStore.compactMode;
-    config.sidebar_docked =
-      typeof cfg.sidebar_docked === "boolean"
-        ? cfg.sidebar_docked
-        : sidebarStore.sidebarDocked;
-    config.sidebar_pinned =
-      typeof cfg.sidebar_pinned === "boolean"
-        ? cfg.sidebar_pinned
-        : sidebarStore.sidebarPinned;
-    config.date_format = userPrefsStore.dateFormat;
-    config.theme_mode = userPrefsStore.themeMode;
-    config.stack_strictness =
-      cfg.stack_strictness != null
-        ? cfg.stack_strictness
-        : config.stack_strictness;
-    const similarityValue =
-      cfg.similarity_character ?? cfg.selected_similarity_character;
-    sortStore.selectedSimilarityCharacter =
-      similarityValue ?? sortStore.selectedSimilarityCharacter ?? null;
-    const newHiddenTags = Array.isArray(cfg.hidden_tags) ? cfg.hidden_tags : [];
-    if (
-      userPrefsStore.hiddenTags.length !== newHiddenTags.length ||
-      userPrefsStore.hiddenTags.some((tag, i) => tag !== newHiddenTags[i])
-    ) {
-      userPrefsStore.hiddenTags = newHiddenTags;
-    }
-    userPrefsStore.applyTagFilter = Boolean(cfg.apply_tag_filter);
-    const rawPt = cfg.smart_score_penalised_tags;
-    if (rawPt && typeof rawPt === "object" && !Array.isArray(rawPt)) {
-      userPrefsStore.penalisedTagWeights = Object.fromEntries(
-        Object.entries(rawPt).map(([k, v]) => [
-          String(k).trim().toLowerCase(),
-          Number(v) || 0,
-        ]),
-      );
-    } else if (Array.isArray(rawPt)) {
-      userPrefsStore.penalisedTagWeights = Object.fromEntries(
-        rawPt.map((t) => [
-          String(t || "")
-            .trim()
-            .toLowerCase(),
-          3,
-        ]),
-      );
-    } else {
-      userPrefsStore.penalisedTagWeights = {};
-    }
-    config.selectedSimilarityCharacter = sortStore.selectedSimilarityCharacter;
-    configSnapshot.value = {
-      sort: sortStore.selectedSort || "",
-      descending: sortStore.selectedDescending,
-      columns: typeof gridStore.columns === "number" ? gridStore.columns : null,
-      sidebar_thumbnail_size:
-        typeof userPrefsStore.sidebarThumbnailSize === "number"
-          ? userPrefsStore.sidebarThumbnailSize
-          : null,
-      show_keyboard_hint: userPrefsStore.showKeyboardHint,
-      show_face_bboxes: gridStore.showFaceBboxes,
-      show_problem_icon: gridStore.showProblemIcon,
-      expand_all_stacks: gridStore.showStacks,
-      compact_mode: gridStore.compactMode,
-      sidebar_docked: sidebarStore.sidebarDocked,
-      sidebar_pinned: sidebarStore.sidebarPinned,
-      date_format: userPrefsStore.dateFormat,
-      theme_mode: userPrefsStore.themeMode,
-      similarity_character: sortStore.selectedSimilarityCharacter,
-      stack_strictness:
-        cfg.stack_strictness != null ? Number(cfg.stack_strictness) : null,
-      hidden_tags: userPrefsStore.hiddenTags,
-      apply_tag_filter: userPrefsStore.applyTagFilter,
-    };
-    filterStore.comfyuiConfigured = Boolean(cfg?.comfyui_url);
-    if (typeof cfg?.public_url === "string" && cfg.public_url) {
-      userPrefsStore.publicUrl = cfg.public_url;
-    }
-    userPrefsStore.embedWatermark = Boolean(cfg?.embed_watermark);
-    userPrefsStore.hidePurgeSnapshotWarning = Boolean(
-      cfg?.hide_purge_snapshot_warning,
-    );
-    const cfu = cfg?.check_for_updates;
-    userPrefsStore.checkForUpdates =
-      cfu === true ? true : cfu === false ? false : null;
-    if (userPrefsStore.checkForUpdates === null) {
-      updateCheckDialogOpen.value = true;
-    }
-  } catch (e) {
-    console.error("Failed to fetch user config:", e);
-  } finally {
-    configApplying.value = false;
-    configLoading.value = false;
-    configLoaded.value = true;
-  }
-}
-
-async function patchConfigUIOptions() {
-  if (!configLoaded.value || configLoading.value || configApplying.value)
-    return;
-  const patch = {};
-  if (sortStore.selectedSort) patch.sort = sortStore.selectedSort;
-  patch.descending = sortStore.selectedDescending;
-  patch.thumbnail_size_level = gridStore.sizeLevel;
-  // Keep the legacy `columns` field in sync with the derived count so older
-  // clients (and anything still reading `columns`) stay consistent.
-  if (gridStore.columns) patch.columns = gridStore.columns;
-  if (userPrefsStore.sidebarThumbnailSize) {
-    patch.sidebar_thumbnail_size = userPrefsStore.sidebarThumbnailSize;
-  }
-  if (userPrefsStore.sidebarWidth) {
-    patch.sidebar_width = userPrefsStore.sidebarWidth;
-  }
-  if (typeof userPrefsStore.showKeyboardHint === "boolean")
-    patch.show_keyboard_hint = userPrefsStore.showKeyboardHint;
-  if (typeof gridStore.showFaceBboxes === "boolean") {
-    patch.show_face_bboxes = gridStore.showFaceBboxes;
-  }
-  if (typeof gridStore.showProblemIcon === "boolean") {
-    patch.show_problem_icon = gridStore.showProblemIcon;
-  }
-  if (typeof gridStore.showStacks === "boolean") {
-    patch.expand_all_stacks = gridStore.showStacks;
-  }
-  if (typeof gridStore.compactMode === "boolean") {
-    patch.compact_mode = gridStore.compactMode;
-  }
-  if (
-    gridStore.thumbnailMode === "square" ||
-    gridStore.thumbnailMode === "justified"
-  ) {
-    patch.thumbnail_mode = gridStore.thumbnailMode;
-  }
-  if (typeof sidebarStore.sidebarDocked === "boolean") {
-    patch.sidebar_docked = sidebarStore.sidebarDocked;
-  }
-  if (typeof sidebarStore.sidebarPinned === "boolean") {
-    patch.sidebar_pinned = sidebarStore.sidebarPinned;
-  }
-  if (
-    typeof userPrefsStore.dateFormat === "string" &&
-    userPrefsStore.dateFormat
-  ) {
-    patch.date_format = userPrefsStore.dateFormat;
-  }
-  if (
-    typeof userPrefsStore.themeMode === "string" &&
-    userPrefsStore.themeMode
-  ) {
-    patch.theme_mode = userPrefsStore.themeMode;
-  }
-  if (sortStore.selectedSimilarityCharacter != null) {
-    patch.similarity_character = sortStore.selectedSimilarityCharacter;
-  }
-  if (sortStore.stackThreshold != null && sortStore.stackThreshold !== "") {
-    const parsed = parseFloat(String(sortStore.stackThreshold));
-    if (Number.isFinite(parsed)) {
-      patch.stack_strictness = parsed;
-    }
-  }
-
-  const snapshot = configSnapshot.value || {};
-  const changed = Object.fromEntries(
-    Object.entries(patch).filter(([key, value]) => snapshot[key] !== value),
-  );
-  if (Object.keys(changed).length === 0) {
-    return;
-  }
-
-  try {
-    await patchUserConfig(changed);
-    configSnapshot.value = { ...snapshot, ...changed };
-  } catch (e) {
-    console.error("Error patching user config:", e);
-  }
-}
-
 /**
  * Is a modal surface (a dialog, the lightbox) currently covering the app?
  *
@@ -1766,16 +1484,6 @@ watch([() => searchStore.searchInput, () => searchStore.searchHistory], () => {
 });
 
 watch(
-  [() => sortStore.selectedSort, () => sortStore.selectedDescending],
-  () => {
-    patchConfigUIOptions();
-    // No refreshGridVersion() here: ImageGrid's selection watch already fires
-    // when selectedSort/selectedDescending props change, so calling
-    // refreshGridVersion() would produce a redundant second grid fetch.
-  },
-);
-
-watch(
   () => userPrefsStore.hiddenTags,
   () => {
     gridStore.refreshGridVersion();
@@ -1806,116 +1514,6 @@ watch(
 );
 
 watch(
-  () => gridStore.thumbnailSize,
-  () => {
-    patchConfigUIOptions();
-    updateMaxColumns();
-  },
-);
-
-watch(
-  () => userPrefsStore.showKeyboardHint,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  [
-    () => gridStore.showFaceBboxes,
-    () => gridStore.showProblemIcon,
-    () => gridStore.showStacks,
-    () => gridStore.compactMode,
-  ],
-  () => {
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  [
-    () => gridStore.showFaceBboxes,
-    () => gridStore.showProblemIcon,
-    () => gridStore.showStacks,
-  ],
-  () => {},
-  { immediate: true },
-);
-
-watch(
-  () => sortStore.selectedSimilarityCharacter,
-  () => {
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => sortStore.stackThreshold,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => gridStore.sizeLevel,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => gridStore.thumbnailMode,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => sidebarStore.sidebarDocked,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => sidebarStore.sidebarPinned,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => userPrefsStore.sidebarThumbnailSize,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => userPrefsStore.sidebarWidth,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-  },
-);
-
-watch(
-  () => userPrefsStore.dateFormat,
-  () => {
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
-    gridStore.refreshGridVersion();
-  },
-);
-
-watch(
   () => gridStore.gridVersion,
   () => {
     wsStore.clearPendingExternalImportIds();
@@ -1927,8 +1525,6 @@ watch(
   () => userPrefsStore.themeMode,
   (value) => {
     theme.global.name.value = resolveThemeName(value);
-    if (!configLoaded.value) return;
-    patchConfigUIOptions();
   },
   { immediate: true },
 );
